+++ packages/tiptap-extensions/web/src/styles.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames';
 import { camelCase, upperFirst } from 'lodash';
 import type { CSSProperties } from 'react';
 import React from 'react';
-import type { RicosExtension } from 'ricos-types';
+import type { RicosExtension, RicosExtensionConfig } from 'ricos-types';
 import type { RichContentTheme } from 'wix-rich-content-common';
 import generalRTLIgnoredStyles from 'wix-rich-content-common/dist/statics/styles/general.rtlignore.scss';
 import generalStyles from 'wix-rich-content-editor-common/dist/statics/styles/general.scss';
@@ -92,18 +92,21 @@ export const styles: RicosExtension = {
   type: 'extension' as const,
   groups: [],
   name: 'style',
+  reconfigure: (config: RicosExtensionConfig, extensions: RicosExtension[]) => ({
+    ...config,
+    addOptions: () => ({}),
+    addNodeHoc: () => ({
+      nodeTypes: extensions
+        .filter(extension => !extension.groups.includes('custom-styles'))
+        .map(({ name }) => name),
+      priority: 100,
+      nodeHoc: StylesHOC,
+    }),
+  }),
   createExtensionConfig() {
     return {
       name: this.name,
       priority: 30,
-      addOptions: () => ({}),
-      addNodeHoc() {
-        return {
-          nodeTypes: ['*'],
-          nodeHoc: StylesHOC,
-          priority: 100,
-        };
-      },
     };
   },
 };